+++ src/main/java/ru/practicum/shareit/item/ItemService.java
@@ -3,13 +3,13 @@ package ru.practicum.shareit.item;
 import java.util.List;
 
 public interface ItemService {
-    ItemDto createItem(ItemDto itemDto, Integer ownerId);
+    ItemDto createItem(ItemDto itemDto, Long ownerId);
 
-    ItemDto getItem(Integer itemId);
+    ItemDto getItem(Long itemId);
 
-    ItemDto updateItem(ItemDto itemDto, Integer itemId, Integer ownerId);
+    ItemDto updateItem(ItemDto itemDto, Long itemId, Long ownerId);
 
-    List<ItemDto> getItemsByOwnerId(Integer ownerId);
+    List<ItemDto> getItemsByOwnerId(Long ownerId);
 
     List<ItemDto> getSearchedItems(String searchString);
 }
